all comments addressed
